+++ snail-race-bet/java/src/main/java/snail/race/kata/domain/BetRepository.java
@@ -6,5 +6,5 @@ import java.util.List;
 public interface BetRepository {
     void register( Bet bet);
 
-    List<Bet> findByDateRange(int from, int to);
+    List<Bet> findByDateRange(long from, long to);
 }
